Add deferred resource release for D3D12 objects
Introduced DeferredReleaseResource template and deferred release logic in ModuleD3D12 to safely release D3D12 resources after GPU usage is complete. Updated RenderTexture and related modules to use deferred release, replaced unsigned frame indices with UINT64 for correctness, and added memory tracking for deferred releases.

+++ Source/ModuleResources.cpp
@@ -247,9 +247,7 @@ ComPtr<ID3D12Resource> ModuleResources::getUploadHeap(size_t size)
 
 void ModuleResources::preRender()
 {
-    // collect garbage
-
-    UINT completedFrame = app->getD3D12()->getLastCompletedFrame();
+    UINT64 completedFrame = app->getD3D12()->getLastCompletedFrame();
 
     for (int i = 0; i < deferredFrees.size(); ++i)
     {
